+++ lib/map_location_picker.dart
@@ -1,9 +1,11 @@
 library map_location_picker;
 
-export 'src/map_location_picker.dart';
-export 'src/autocomplete_service.dart';
-export 'src/autocomplete_view.dart';
-export 'package:google_maps_webservice/places.dart';
-export 'package:google_maps_webservice/geocoding.dart';
 export 'package:geolocator/geolocator.dart';
 export 'package:google_maps_flutter/google_maps_flutter.dart';
+export 'package:google_maps_webservice/geocoding.dart';
+export 'package:google_maps_webservice/places.dart';
+
+export 'src/autocomplete_service.dart';
+export 'src/autocomplete_view.dart';
+export 'src/map_location_picker.dart';
+export 'src/utils.dart';